# packages/apigateway/openapi.yaml
openapi: 3.1.0
info:
  title: Desirelines API
  description: |
    REST API for serving Strava activity data to the web frontend.

    ## Authentication
    All `/v1/activities` endpoints require Firebase authentication.
    Include `Authorization: Bearer <firebase-token>` header.

    ## Pagination
    The `/v1/activities` list endpoint uses cursor-based pagination.
    Use `next_cursor` from the response to fetch subsequent pages.

    ## Versioning policy
    `/v1/` is **additive-only**: new optional response fields, new endpoints,
    and new optional query params can ship at any time. Existing fields are
    not removed, retyped, or semantically repurposed in place — if a field's
    meaning needs to change, add a new field. The frontend's Zod contracts
    use `.passthrough()` so new fields don't trigger validation errors.

    A truly breaking change (which should be rare on a small app) means
    introducing a new endpoint or, only if there's no other option, a `/v2/`
    surface alongside `/v1/`. No formal sunset policy — the frontend is the
    sole consumer and migrates in lockstep.
  version: 1.0.0
  contact:
    name: Desirelines
    url: https://github.com/andy-esch/desirelines
servers:
  - url: https://desirelines.andyes.ch/api
    description: Production (via Firebase Hosting proxy)
  - url: http://localhost:8080/api
    description: Local development
...
paths:
...
  /v1/activities/routes/tiles/{z}/{x}/{y}:
    get:
      tags: [Activities]
      summary: Activity routes as a Mapbox Vector Tile
      description: >
        Returns the user's real-world activity routes for the given tile as a
        Mapbox Vector Tile (ST_AsMVT). Only geo-bearing activities (those tagged
        to >=1 region) are included; virtual/indoor activities are excluded. An
        empty-but-valid tile is returned when the tile has no features.
      operationId: getRouteTile
      security:
        - bearerAuth: []
      parameters:
        - name: z
          in: path
          required: true
          description: Tile zoom level (0-22)
          schema:
            type: integer
            minimum: 0
            maximum: 22
        - name: x
          in: path
          required: true
          description: Tile column
          schema:
            type: integer
            minimum: 0
        - name: y
          in: path
          required: true
          description: Tile row
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: A Mapbox Vector Tile (protobuf)
          content:
            application/vnd.mapbox-vector-tile:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase JWT
      description: Firebase ID token
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    ServiceUnavailable:
      description: Database unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      description: |
        Standardized error response envelope returned by all error paths.
        `error` is the only required field; the rest are added when
        available (request ID middleware, active OTel span, structured
        details). All keys are camelCase to match success responses.
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
          description: Human-readable error message.
          example: Invalid year format
        code:
          type: string
          description: Machine-readable error code (see apierrors taxonomy).
          example: BAD_REQUEST
        requestId:
          type: string
          description: Server-generated request ID for support correlation.
          example: 01J0X8C0F5H9TZ3K9R8VW3K2QH
        traceId:
          type: string
          description: |
            Raw hex W3C trace ID of the active OTel span (32 lowercase
            hex chars). Omitted when no span is active.
          example: 0af7651916cd43dd8448eb211c80319c
        details:
          type: object
          description: |
            Optional structured context for the error (e.g. validation
            field hints, retry timing). Shape varies by error site; do
            not depend on specific keys.
          additionalProperties: true
...